+++ include/glaze/binary/write.hpp
@@ -11,7 +11,7 @@
 #include "glaze/core/reflection_tuple.hpp"
 #include "glaze/core/seek.hpp"
 #include "glaze/core/write.hpp"
-#include "glaze/reflection/reflect.hpp"
+#include "glaze/core/refl.hpp"
 #include "glaze/util/dump.hpp"
 #include "glaze/util/for_each.hpp"
 #include "glaze/util/variant.hpp"
