+++ include/glaze/core/reflection_tuple.hpp
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "glaze/core/common.hpp"
-#include "glaze/reflection/reflect.hpp"
+#include "glaze/core/refl.hpp"
 
 namespace glz::detail
 {
