+++ include/glaze/mustache/mustache.hpp
@@ -5,7 +5,7 @@
 
 #include "glaze/core/read.hpp"
 #include "glaze/core/write.hpp"
-#include "glaze/reflection/reflect.hpp"
+#include "glaze/core/refl.hpp"
 
 namespace glz
 {
